+++ incs/mini_header.h
@@ -6,7 +6,7 @@
 /*   By: dgarcez- <dgarcez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/03/11 16:55:53 by dpaes-so          #+#    #+#             */
-/*   Updated: 2025/04/21 15:34:16 by dgarcez-         ###   ########.fr       */
+/*   Updated: 2025/04/22 19:34:41 by dgarcez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -95,10 +95,15 @@ t_tokentype				token_type(char *value);
 bool					is_token(char **input, int *len, int *flag);
 bool					is_quote(char **input, int *len);
 bool					skip_wspaces(char **input);
+bool					skip_fake_wspaces(char *input);
 bool					check_next(char *input);
 bool					check_redir(t_token tokens);
 bool					error_syntax(t_token *tokens);
 int						count_nodes(t_token *tokens);
-void					create_array(t_token *tokens);
+t_token					**create_array(t_token *tokens);
+void					create_tree(t_tree **tree_root, t_token **array, bool pipe, int *i);
+void					tree_apply_infix(t_tree *root, int level);
+void					free_tree(t_tree *root);
+void					free_array(t_token **array);
 
 #endif
