+++ srcs/parser/ast_tree.c
@@ -6,200 +6,37 @@
 /*   By: dgarcez- <dgarcez-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/17 15:14:20 by dgarcez-          #+#    #+#             */
-/*   Updated: 2025/04/21 20:34:57 by dgarcez-         ###   ########.fr       */
+/*   Updated: 2025/04/22 20:03:14 by dgarcez-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incs/mini_header.h"
 
-// void	put_redir(t_token *redirs, t_token *tokens, int pos)
-// {
-// 	int	i;
-// 	int	j;
-
-// 	j = 0;
-// 	i = 0;
-// 	while(i < pos)
-// 	{
-// 		if (tokens[i].type >= T_HERE_DOC || tokens[i].type <= T_APPEND_REDIR)
-// 		{
-// 			redirs[j].type = tokens[i].type;
-// 			redirs[j].value = ft_strdup(tokens[i].value);
-// 			j++;
-// 		}
-// 		i++;
-// 	}
-// }
-
-// t_token	*get_redirs(t_token *tokens, int pos)
-// {
-// 	int		i;
-// 	int		count;
-// 	t_token	*redirs;
-
-// 	count = 0;
-// 	i = 0;
-// 	while (i <= pos)
-// 	{
-// 		if (tokens[i].type >= T_HERE_DOC || tokens[i].type <= T_APPEND_REDIR)
-// 			count++;
-// 		i++;
-// 	}
-// 	if (count == 0)
-// 		return (NULL);
-// 	redirs = ft_calloc(count + 1, sizeof(t_token));
-// 	if (redirs == NULL)
-// 		return (NULL);
-// 	redirs = put_redir(redirs, tokens, pos);
-// 	return (redirs);
-// }
-
-// t_tree	*create_node(t_token *tokens, int pos)
-// {
-// 	t_tree	*ast;
-
-// 	ast = malloc(sizeof(t_tree));
-// 	if (ast == NULL)
-// 		return (NULL);
-// 	ast->left = NULL;
-// 	ast->right = NULL;
-// 	if (pos == 0)
-// 		ast->node.cmd = token.value;
-// 	else
-// 		ast->node.cmd = NULL;
-// 	ast->node.redirections = get_redirs(tokens, pos);
-// 		if (token.type == T_PIPE)
-// 			ast->node.pipe = true;
-// 	return (node);
-// }
-
-int	count_elems(t_token *tokens)
-{
-	int	i;
-	int	count;
-
-	count = 0;
-	i = 0;
-	while (tokens[i].type != T_NULL)
-	{
-		if (tokens[i].type != T_PIPE)
-			count++;
-		while (tokens[i].type != T_NULL && tokens[i].type != T_PIPE)
-			i++;
-		if (tokens[i].type != T_NULL && tokens[i].type == T_PIPE)
-		{
-			count++;
-			i++;
-		}
-	}
-	return (count);
-}
-
-int	token_count(t_token *tokens, int *i)
-{
-	int	amount;
-
-	amount = 0;
-	if (tokens[*i].type == T_PIPE)
-	{
-		(*i)++;
-		return (1);
-	}
-	else
-	{
-		while (tokens[*i].type != T_NULL && tokens[*i].type != T_PIPE)
-		{
-			(*i)++;
-			amount++;
-		}
-	}
-	return (amount);
-}
-
-t_token	**array_creation(t_token *tokens)
-{
-	int		i;
-	int		index;
-	int		len;
-	int		amount;
-	t_token	**array;
-
-	i = 0;
-	index = 0;
-	amount = count_elems(tokens);
-	array = ft_calloc(amount + 1, sizeof(t_token *));
-	if (array == NULL)
-		return (NULL);
-	array[amount] = NULL;
-	while (i < amount)
-	{
-		len = token_count(tokens, &index);
-		array[i] = ft_calloc(len + 1, sizeof(t_token));
-		if (array[i] == NULL)
-			return (NULL);
-		array[i][len].value = NULL;
-		array[i][len].type = T_NULL;
-		i++;
-	}
-	return (array);
-}
-
-void	init_array(t_token **array, t_token *tokens)
-{
-	int	i;
-	int	j;
-	int	k;
-
-	i = -1;
-	k = 0;
-	while (array[++i])
-	{
-		j = 0;
-		if (array[i][j].type != T_NULL && tokens[k].type == T_PIPE)
-		{
-			array[i][j].value = ft_strdup(tokens[k++].value);
-			array[i]->type = T_PIPE;
-			j++;
-		}
-		else
-		{
-			while (array[i][j].type != T_NULL && array[i][j].type != T_PIPE)
-			{
-				array[i][j].value = ft_strdup(tokens[k].value);
-				array[i][j].type = tokens[k++].type;
-				j++;
-			}
-		}
-	}
-}
-
-void	free_array(t_token **array)
-{
-	int	i;
-	int	j;
-
-	i = 0;
-	while (array[i])
-	{
-		j = 0;
-		while (array[i][j].type != T_NULL)
-		{
-			free(array[i][j].value);
-			j++;
-		}
-		free(array[i]);
-		i++;
-	}
-	free(array);
-}
-
 void	init_redirs(t_tree *tree_node, t_token *tokens)
 {
 	int	i;
 	int	j;
+
+	i = 0;
+	j = 0;
+	while (tokens[i].type != T_NULL)
+	{
+		if (tokens[i].type >= T_HERE_DOC && tokens[i].type <= T_APPEND_REDIR)
+		{
+			tree_node->node.redirections[j].value = ft_strdup(tokens[i].value);
+			tree_node->node.redirections[j].type = tokens[i].type;
+			// skip_wspaces(&tree_node->node.redirections[j].value);
+			j++;
+		}
+		i++;
+	}
+}
+
+void	make_redirs(t_tree *tree_node, t_token *tokens)
+{
+	int	i;
 	int	count;
 
-	j = 0;
 	i = 0;
 	count = 0;
 	while (tokens[i].type != T_NULL)
@@ -213,17 +50,7 @@ void	init_redirs(t_tree *tree_node, t_token *tokens)
 		return ;
 	tree_node->node.redirections[count].value = NULL;
 	tree_node->node.redirections[count].type = T_NULL;
-	i = 0;
-	while (tokens[i].type != T_NULL)
-	{
-		if (tokens[i].type >= T_HERE_DOC && tokens[i].type <= T_APPEND_REDIR)
-		{
-			tree_node->node.redirections[j].value = ft_strdup(tokens[i].value);
-			tree_node->node.redirections[j].type = tokens[i].type;
-			j++;
-		}
-		i++;
-	}
+	init_redirs(tree_node, tokens);
 }
 
 void	init_tree_node(t_tree *tree_node, t_token *tokens)
@@ -231,7 +58,7 @@ void	init_tree_node(t_tree *tree_node, t_token *tokens)
 	int	i;
 
 	i = 0;
-	init_redirs(tree_node, tokens);
+	make_redirs(tree_node, tokens);
 	while (tokens[i].type != T_NULL && tokens[i].type >= T_HERE_DOC
 		&& tokens[i].type <= T_APPEND_REDIR)
 		i++;
@@ -242,11 +69,6 @@ void	init_tree_node(t_tree *tree_node, t_token *tokens)
 			return ;
 		i++;
 	}
-	// else
-	// {
-	// 	tree_node->node.cmd = NULL;
-	// 	tree_node->node.args = NULL;
-	// }
 	while (tokens[i].type != T_NULL)
 	{
 		if (tokens[i].type == T_WORD)
@@ -290,9 +112,9 @@ void	create_tree(t_tree **tree_root, t_token **array, bool pipe, int *i)
 		{
 			*tree_root = tree_create_node(array[*i], 1);
 			*i -= 1;
-			create_tree(&((*tree_root)->left), array, true, i);
+			create_tree(&((*tree_root)->right), array, true, i);
 			*i += 2;
-			create_tree(&((*tree_root)->right), array, false, i);
+			create_tree(&((*tree_root)->left), array, false, i);
 			return ;
 		}
 		else
@@ -305,104 +127,3 @@ void	create_tree(t_tree **tree_root, t_token **array, bool pipe, int *i)
 	}
 }
 
-void	print_tabs(int level)
-{
-	while (level != 0)
-	{
-		printf("\t");
-		level--;
-	}
-}
-
-void	print_tree(t_tree *root, int level)
-{
-	int	i;
-
-	i = 0;
-	if (level == 0)
-		printf("ROOT!\n");
-	if (root->node.cmd)
-	{
-		print_tabs(level);
-		printf("cmd = %s\n", root->node.cmd);
-	}
-	if (root->node.args)
-	{
-		print_tabs(level);
-		printf("args = %s\n", root->node.args);
-	}
-	if (root->node.pipe == true)
-	{
-		print_tabs(level);
-		printf("|\n");
-	}
-	while (root->node.redirections != NULL
-		&& root->node.redirections[i].type != T_NULL)
-	{
-		print_tabs(level);
-		printf("redir = %s type = %u\n", root->node.redirections[i].value,
-				root->node.redirections[i].type);
-		i++;
-	}
-}
-
-void	tree_apply_infix(t_tree *root, int level)
-{
-	if (root->left != NULL)
-		tree_apply_infix(root->left, level + 1);
-	print_tree(root, level);
-	if (root->right != NULL)
-		tree_apply_infix(root->right, level + 1);
-}
-
-void	free_tree(t_tree *root)
-{
-	int	i;
-
-	i = 0;
-	if (!root)
-		return ;
-	if (root->node.args)
-		free(root->node.args);
-	if (root->node.cmd)
-		free(root->node.cmd);
-	while (root->node.redirections != NULL && root->node.redirections[i].value)
-	{
-		free(root->node.redirections[i].value);
-		i++;
-	}
-	free(root->node.redirections);
-	free_tree(root->left);
-	free_tree(root->right);
-	free(root);
-}
-
-void	create_array(t_token *tokens)
-{
-	int		i;
-	int		j;
-	t_token	**array;
-	t_tree	*root;
-
-	root = NULL;
-	i = 0;
-	j = 0;
-	array = array_creation(tokens);
-	init_array(array, tokens);
-	// while (array[i])
-	// {
-	// 	j = 0;
-	// 	while (array[i][j].type != T_NULL)
-	// 	{
-	// 		printf("i = %d j = %d value = %s$ type = %d\n", i, j,
-	// 				array[i][j].value, array[i][j].type);
-	// 		j++;
-	// 	}
-	// 	i++;
-	// }
-	// i = 0;
-	create_tree(&root, array, false, &i);
-	tree_apply_infix(root, 0);
-	free_tree(root);
-	free_array(array);
-}
